refactor: Race 클래스 역할 분담
- 기존 Race 클래스가 domain과 view의 역할을 함께 하고 있었는데 이를 App 클래스를 만들어 역할을 분리
